excise setlib
