+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include versioneer.py
 include src/tomato/_version.py
+include src/tomato/data/*
