--- src/tomato/setlib/__init__.py
@@ -1,1 +1,0 @@
-from .functions import get_settings, get_pipelines, get_dirs
